Move async iterators to ND4J for use in SameDiff
