+++ deeplearning4j/deeplearning4j-core/src/test/java/org/deeplearning4j/datasets/datavec/RecordReaderDataSetiteratorTest.java
@@ -40,7 +40,7 @@ import org.deeplearning4j.BaseDL4JTest;
 import org.deeplearning4j.TestUtils;
 import org.deeplearning4j.datasets.datavec.exception.ZeroLengthSequenceException;
 import org.deeplearning4j.datasets.datavec.tools.SpecialImageRecordReader;
-import org.deeplearning4j.datasets.iterator.AsyncDataSetIterator;
+import org.nd4j.linalg.dataset.AsyncDataSetIterator;;
 import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
