+++ deeplearning4j/deeplearning4j-data/deeplearning4j-utility-iterators/src/main/java/org/deeplearning4j/datasets/iterator/callbacks/DataSetCallback.java
@@ -21,13 +21,9 @@ import org.nd4j.linalg.dataset.api.DataSet;
 import org.nd4j.linalg.dataset.api.MultiDataSet;
 
 /**
- * @author raver119@gmail.com
+ * @deprecated Use {@link org.nd4j.linalg.dataset.callbacks.DataSetCallback}
  */
-public interface DataSetCallback {
+@Deprecated
+public interface DataSetCallback extends org.nd4j.linalg.dataset.callbacks.DataSetCallback {
 
-    void call(DataSet dataSet);
-
-    void call(MultiDataSet multiDataSet);
-
-    void reset();
 }
