+++ deeplearning4j/deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/iterator/SparkADSI.java
@@ -19,7 +19,7 @@ package org.deeplearning4j.spark.iterator;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.TaskContext;
 import org.apache.spark.TaskContextHelper;
-import org.deeplearning4j.datasets.iterator.AsyncDataSetIterator;
+import org.nd4j.linalg.dataset.AsyncDataSetIterator;;
 import org.deeplearning4j.datasets.iterator.callbacks.DataSetCallback;
 import org.deeplearning4j.datasets.iterator.callbacks.DefaultCallback;
 import org.nd4j.linalg.api.memory.MemoryWorkspace;
